refactor: extract bridge_e2e_hooks, llm_request_dump, history_apply, edge_ledger to astra-turn-core (Phase 19)
Move 4 modules (~1,237 LOC) from runtime to astra-turn-core:
- bridge_e2e_hooks.rs: E2E test hooks for bridge testing
- llm_request_dump.rs: LLM request payload dump for post-mortem debugging
- history_apply.rs: Turn inputs to history application
- edge_ledger.rs: Edge callback ledger with poll/take semantics

Added axum dependency to astra-turn-core for bridge_e2e_hooks HeaderMap usage.
Runtime originals replaced with re-exports.

Turn-core: 1627 tests passing
Runtime: 3180 tests + 5 pre-existing remote skill failures
